Made new file for AcademicSystem and renamed react component
